Simplify Setup Page complexity.
