--- src/Onboarding/Setup/PageStatus.php
@@ -1,16 +1,0 @@
-<?php
-
-/**
- * Onboarding class
- *
- * @package Give
- */
-
-namespace Give\Onboarding\Setup;
-
-defined( 'ABSPATH' ) || exit;
-
-class PageStatus {
-	const ENABLED  = 'enabled';
-	const DISABLED = 'disabled';
-}
